+++ app/handlers/__init__.py
@@ -1,0 +1,3 @@
+from . import balance
+
+__all__ = ["balance"]
