+++ app/handlers/balance/__init__.py
@@ -1,0 +1,3 @@
+from .main import register_balance_handlers
+
+__all__ = ["register_balance_handlers"]
